renamed o publicSchema
